requirements.txt
mcp>=1.0.0
docker>=7.0.0
python-dotenv>=1.0.0